Replace adjustBreakpoints with controller.breakpoints.resolver and controller.finished with controller.current.finished
