+++ packages/truffle-debugger/lib/controller/selectors/index.js
@@ -1,11 +1,13 @@
 import debugModule from "debug";
-const debug = debugModule("debugger:controller:sagas"); //eslint-disable-line no-unused-vars
+const debug = debugModule("debugger:controller:selectors"); //eslint-disable-line no-unused-vars
 
 import { createSelectorTree, createLeaf } from "reselect-tree";
 
 import evm from "lib/evm/selectors";
 import solidity from "lib/solidity/selectors";
 import trace from "lib/trace/selectors";
+
+import { anyNonSkippedInRange } from "lib/ast/map";
 
 /**
  * @private
@@ -62,23 +64,72 @@ const controller = createSelectorTree({
        * controller.current.location.isMultiline
        */
       isMultiline: createLeaf([solidity.current.isMultiline], identity)
-    }
+    },
+
+    /**
+     * controller.current.finished
+     */
+    finished: createLeaf([trace.finished], identity)
   },
 
   /**
-   * controller.breakpoints
+   * controller.breakpoints (namespace)
    */
-  breakpoints: createLeaf(["./state"], state => state.breakpoints),
+  breakpoints: {
+    /**
+     * controller.breakpoints (selector)
+     */
+    _: createLeaf(["/state"], state => state.breakpoints),
+
+    /**
+     * controller.breakpoints.resolver (selector)
+     * this selector returns a function that adjusts a given line-based
+     * breakpoint (on node-based breakpoints it simply returns the input) by
+     * repeatedly moving it down a line until it lands on a line where there's
+     * actually somewhere to break.  if no such line exists beyond that point, it
+     * returns null instead.
+     */
+    resolver: createLeaf([solidity.info.sources], sources => breakpoint => {
+      let adjustedBreakpoint;
+      if (breakpoint.node === undefined) {
+        let line = breakpoint.line;
+        let { source, ast } = sources[breakpoint.sourceId];
+        let lineLengths = source.split("\n").map(line => line.length);
+        //why does neither JS nor lodash have a scan function like Haskell??
+        //guess we'll have to do our scan manually
+        let lineStarts = [0];
+        for (let length of lineLengths) {
+          lineStarts.push(lineStarts[lineStarts.length - 1] + length + 1);
+          //+1 for the /n itself
+        }
+        debug(
+          "line: %s",
+          source.slice(lineStarts[line], lineStarts[line] + lineLengths[line])
+        );
+        while (
+          line < lineLengths.length &&
+          !anyNonSkippedInRange(ast, lineStarts[line], lineLengths[line])
+        ) {
+          debug("incrementing");
+          line++;
+        }
+        if (line >= lineLengths.length) {
+          adjustedBreakpoint = null;
+        } else {
+          adjustedBreakpoint = { ...breakpoint, line };
+        }
+      } else {
+        debug("node-based breakpoint");
+        adjustedBreakpoint = breakpoint;
+      }
+      return adjustedBreakpoint;
+    })
+  },
 
   /**
    * controller.isStepping
    */
-  isStepping: createLeaf(["./state"], state => state.isStepping),
-
-  /**
-   * controller.finished
-   */
-  finished: createLeaf([trace.finished], identity)
+  isStepping: createLeaf(["./state"], state => state.isStepping)
 });
 
 export default controller;
